refactor(terminal): move commands to core-common
